requirements.txt
langchain
langchain-core
langchain-community 
chromadb
sentence-transformers
streamlit
pypdf
